Use path prefix to handle header and allow deployment to github pages
This can also be done by creating a new account and deploying to the root path of that account for free.

In that case, you would make a repository with the name:

MYACCOUNT.github.io (ex. andreiboris.github.io) 

Then, you would associate this project with that repository.

Then, remove prefix path from the gatsby-config.js file. We will not be using a prefix path for this deployment (that is the whole point :P )

To deployment, add a -b master flag (ex. gh-pages -d public -b master) to deploy straight from the master branch.